[REFACTOR] Move service keys to their own module
