+++ components/core/src/crypto/keys/cache.rs
@@ -1,12 +1,12 @@
 use super::ring_key::RingKey;
 use crate::{crypto::{hash,
-                     keys::{box_key_pair::{OriginPublicEncryptionKey,
-                                           ServicePublicEncryptionKey,
-                                           ServiceSecretEncryptionKey},
+                     keys::{box_key_pair::OriginPublicEncryptionKey,
                             sig_key_pair::{PublicOriginSigningKey,
                                            SecretOriginSigningKey},
                             Key,
                             NamedRevision,
+                            ServicePublicEncryptionKey,
+                            ServiceSecretEncryptionKey,
                             UserPublicEncryptionKey,
                             UserSecretEncryptionKey}},
             error::{Error,
