+++ components/core/src/crypto/keys/user_key.rs
@@ -1,10 +1,10 @@
 use crate::{crypto::{keys::{box_key_pair::{EncryptedSecret,
-                                           ServicePublicEncryptionKey,
                                            SignedBox,
                                            WrappedSealedBox},
                             Key,
                             KeyRevision,
-                            NamedRevision},
+                            NamedRevision,
+                            ServicePublicEncryptionKey},
                      PUBLIC_BOX_KEY_VERSION,
                      SECRET_BOX_KEY_VERSION},
             error::{Error,
